+++ jax/experimental/jax2tf/tests/primitive_harness.py
@@ -1202,7 +1202,11 @@ def _make_scatter_harness(name,
               "unimplemented",
               devices="tpu",
               dtypes=np.complex64,
-              enabled=(f_lax in [lax.scatter_max, lax.scatter_min]))
+              enabled=(f_lax in [lax.scatter_max, lax.scatter_min])),
+          Limitation(
+              "unimplemented",
+              dtypes=np.bool_,
+              enabled=(f_lax in [lax.scatter_add, lax.scatter_mul])),
       ],
       f_lax=f_lax,
       shape=shape,
@@ -1219,8 +1223,8 @@ for dtype in jtu.dtypes.all:
   for f_lax in [
       lax.scatter_add, lax.scatter_mul, lax.scatter_max, lax.scatter_min
   ]:
-    if f_lax in [lax.scatter_add, lax.scatter_mul] and dtype == np.bool_:
-      continue
+    #if f_lax in [lax.scatter_add, lax.scatter_mul] and dtype == np.bool_:
+    #  continue
     _make_scatter_harness("dtypes", dtype=dtype, f_lax=f_lax)
 
 # Validate f_lax/update_jaxpr